Crate tf_format

Structs

Angle

AxisAngle

Rotation represented in axis-angle format.

CoordTransform

Represent coordinate transformation in 3D Euclidean space.

Euler

Rotation represented as Euler angles.

EulerAxisOrder

MaybeTransform

Represent either a rigid transformation or a rotation in 3D Euclidean space.

Quaternion

Rotation represented as a quaternion.

Rodrigues

Rotation represented in Rodrigues format.

RotationMatrix

Transform

Represent a rigid transformation in 3D Euclidean space.

TransformSet

Represent a set of related or disjoint coordinate transformations.

Translation

Represent a translation in 3D Euclidean space.

Enums

EulerAxis

Rotation

Generic rotation data type.

Type Aliases

Length